requirements.txt
beautifulsoup4
requests
pytest
Click
Flask
itsdangerous
Jinja2
MarkupSafe
Werkzeug
Flask-Bootstrap
flask-sqlalchemy
pytest-flask